Remove audit on push
This can be added later. However, it's currently making it too difficult
to automate a release. The dependency that's causing this problem is
acorn and the vulnerability is a denial of service when using a specific
unicode regex. This is not applicable to Validate in production.

Audit on push should be re-added in the future when the downstream
audit is fixed. This requires at least jest and eslint to patch their
dependency on acorn.